requirements.txt
email-validator
fabric
fastapi
httpx
Jinja2
pandas
paramiko
postgrest
psycopg2-binary
pydantic
pydantic[email]
pytest
python-dotenv
python-gitlab
python-multipart
SQLAlchemy
starlette
storage3
supabase
uvicorn
